v3-the-agent-loop/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-qdrant>=0.2.0
qdrant-client>=1.12.0
